// src/protocol/types.ts
export interface User {
  AccountID: number;
  Avatar?: string;
  AwaySince?: string;
  BanMute: boolean;
  BanVotes: boolean;
  BanSpecChat: boolean;
  BattleID?: number;
  Clan?: string;
  Country?: string;
  DisplayName?: string;
  Faction?: string;
  InGameSince?: string;
  IsAdmin: boolean;
  IsBot: boolean;
  LobbyVersion?: string;
  Name?: string;
  SteamID?: string;
  Badges?: string[];
  Icon?: string;
  EffectiveMmElo: number;
  EffectiveElo: number;
  Level: number;
  Rank: number;
}

export interface UserDisconnected {
  Name?: string;
  Reason?: string;
}

export interface Say {
  IsEmote: boolean;
  Place: E.SayPlace;
  Ring: boolean;
  Target?: string;
  Text?: string;
  Time?: string;
  User?: string;
  Source?: E.SaySource;
}

export interface OpenBattle {
  Header?: BattleHeader;
}

export interface BattleHeader {
  BattleID?: number;
  Engine?: string;
  Founder?: string;
  Game?: string;
  IsMatchMaker?: boolean;
  IsRunning?: boolean;
  Map?: string;
  MaxPlayers?: number;
  Mode?: E.AutohostMode;
  Password?: string;
  PlayerCount?: number;
  RunningSince?: string;
  SpectatorCount?: number;
  Title?: string;
  TimeQueueEnabled?: boolean;
  MaxEvenPlayers?: number;
}
// ...
export interface UpdateUserBattleStatus {
  AllyNumber?: number;
  IsSpectator?: boolean;
  QueueOrder?: number;
  Name?: string;
  Sync?: E.SyncStatuses;
  JoinTime?: string;
}

export interface UpdateBotStatus {
  AiLib?: string;
  AllyNumber?: number;
  Name?: string;
  Owner?: string;
}